[Feat] [GGFE-144] coin 정보 props로 넘길 수 있게 수정
